+++ R/spread.R
@@ -83,8 +83,8 @@ spread_.data.frame <- function(data, key_col, value_col, fill = NA,
   }
 
   dim(ordered) <- c(attr(row_id, "n"), attr(col_id, "n"))
+  ordered <- as.data.frame.matrix(ordered, stringsAsFactors = FALSE)
   colnames(ordered) <- as.character(col_labels[[1]])
-  ordered <- as.data.frame(ordered, stringsAsFactors = FALSE)
 
   if (convert) {
     ordered[] <- lapply(ordered, type.convert, as.is = TRUE)