+++ NEWS.md
@@ -1,5 +1,8 @@
 # tidyr 0.1.0.9000
 
+* Spread converts factor value variable to character, instead of embedding
+  matrix inside data frame (#35).
+
 * Preserve negative signs in `extract_numeric()` (#20).
 
 * `separate()` gains `extra` argument which lets you control what happens